SOL-151500: Extract shared field-coercion helpers into one file for postprocess aggregation handlers
